requirements.txt
Flask==3.0.2
pyzbar==0.1.9
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
ultralytics==8.1.16